+++ pkg/network/anonymity/logger/log_builder.go
@@ -11,6 +11,7 @@ var (
 
 type sLogBuilder struct {
 	fService string
+	fRecv    bool
 	fType    ILogType
 	fHash    []byte
 	fProof   uint64
@@ -35,6 +36,11 @@ func (p *sLogBuilder) Get() ILogGetter {
 
 func (p *sLogBuilder) WithType(pType ILogType) ILogBuilder {
 	p.fType = pType
+	return p
+}
+
+func (p *sLogBuilder) WithRecv(pRecv bool) ILogBuilder {
+	p.fRecv = pRecv
 	return p
 }