+++ pkg/network/anonymity/logger/logger_test.go
@@ -11,6 +11,7 @@ import (
 
 const (
 	tcService = "ServiceName"
+	tcRecv    = true
 	tcHash    = "hash-example"
 	tcProof   = 3
 	tcSize    = 8192
@@ -22,6 +23,7 @@ func TestLogger(t *testing.T) {
 	pubKey := asymmetric.LoadRSAPubKey(testutils.TgPubKeys[0])
 	builder := NewLogBuilder(tcService).
 		WithConn(nil).
+		WithRecv(tcRecv).
 		WithHash([]byte(tcHash)).
 		WithProof(tcProof).
 		WithPubKey(pubKey).
@@ -31,6 +33,11 @@ func TestLogger(t *testing.T) {
 	getter := builder.Get()
 	if getter.GetService() != tcService {
 		t.Error("getter.GetService() != tcService")
+		return
+	}
+
+	if getter.GetRecv() != tcRecv {
+		t.Error("getter.GetRecv() != tcRecv")
 		return
 	}