+++ pkg/network/anonymity/logger/types.go
@@ -10,13 +10,12 @@ type (
 )
 
 const (
-	CLogFinal = CLogErroEncryptPayload
+	CLogFinal = CLogErroDatabaseSet
 )
 
 const (
 	// Base (can be in >1 more state)
 	CLogBaseBroadcast ILogType = iota + 1
-	CLogBaseMessageType
 	CLogBaseEnqueueRequest
 	CLogBaseEnqueueResponse
 	CLogBaseGetResponse
@@ -28,21 +27,23 @@ const (
 	CLogInfoWithoutResponse
 
 	// WARN
+	CLogWarnMessageType
 	CLogWarnMessageNull
 	CLogWarnNotFriend
 	CLogWarnUnknownRoute
 	CLogWarnIncorrectResponse
 
 	// ERRO
+	CLogErroEncryptPayload
 	CLogErroDatabaseGet
 	CLogErroDatabaseSet
-	CLogErroEncryptPayload
 )
 
 type ILogBuilder interface {
 	ILogGetterFactory
 
 	WithType(ILogType) ILogBuilder
+	WithRecv(bool) ILogBuilder
 	WithSize(int) ILogBuilder
 	WithProof(uint64) ILogBuilder
 	WithHash([]byte) ILogBuilder
@@ -56,6 +57,7 @@ type ILogGetterFactory interface {
 
 type ILogGetter interface {
 	GetService() string
+	GetRecv() bool
 	GetType() ILogType
 	GetSize() uint64
 	GetProof() uint64